@@ -55,8 +55,8 @@ func (dif *TypedInformerFactory) Get(ctx context.Context, gvr schema.GroupVersio
 
 	listObj := dif.Type.GetListType()
 	lw := &cache.ListWatch{
-		ListFunc:  asStructuredLister(ctx, dif.Client.Resource(gvr).List, listObj),
-		WatchFunc: AsStructuredWatcher(ctx, dif.Client.Resource(gvr).Watch, dif.Type),
+		ListWithContextFunc:  asStructuredLister(dif.Client.Resource(gvr).List, listObj),
+		WatchFuncWithContext: AsStructuredWatcher(dif.Client.Resource(gvr).Watch, dif.Type),
 	}
 	inf := cache.NewSharedIndexInformer(lw, dif.Type, dif.ResyncPeriod, cache.Indexers{
 		cache.NamespaceIndex: cache.MetaNamespaceIndexFunc,
@@ -75,8 +75,8 @@ func (dif *TypedInformerFactory) Get(ctx context.Context, gvr schema.GroupVersio
 
 type unstructuredLister func(context.Context, metav1.ListOptions) (*unstructured.UnstructuredList, error)
 
-func asStructuredLister(ctx context.Context, ulist unstructuredLister, listObj runtime.Object) cache.ListFunc {
-	return func(opts metav1.ListOptions) (runtime.Object, error) {
+func asStructuredLister(ulist unstructuredLister, listObj runtime.Object) cache.ListWithContextFunc {
+	return func(ctx context.Context, opts metav1.ListOptions) (runtime.Object, error) {
 		ul, err := ulist(ctx, opts)
 		if err != nil {
 			return nil, err
@@ -93,8 +93,8 @@ type structuredWatcher func(ctx context.Context, opts metav1.ListOptions) (watch
 
 // AsStructuredWatcher is public for testing only.
 // TODO(mattmoor): Move tests for this to `package duck` and make private.
-func AsStructuredWatcher(ctx context.Context, wf structuredWatcher, obj runtime.Object) cache.WatchFunc {
-	return func(lo metav1.ListOptions) (watch.Interface, error) {
+func AsStructuredWatcher(wf structuredWatcher, obj runtime.Object) cache.WatchFuncWithContext {
+	return func(ctx context.Context, lo metav1.ListOptions) (watch.Interface, error) {
 		uw, err := wf(ctx, lo)
 		if err != nil {
 			return nil, err
